Modify force_mask checking
It seemed to me we can just check if force_mask is available in the metadata and use it. Otherwise we can return false.

However, we would like force_mask to be type stable. So maybe we should always have a vector, which defaults to all forces.
At the minute it returns a vector of length 3, is this correct? There are two atoms in the system so, should it not be 2, or 6 if it's 3N?