Improve the PdH EAM model
Uses NeighbourLists.jl to calculate the potential for periodic systems. Unfortunately, this package doesn't seem to work with any of the automatic differentiation packages I've tried. This means I've had to resort to using FiniteDifferences.jl to calculate the gradient. There is always the option of manually coding the derivative but it seems quite complicated for this model.