Update nrpmd.md
It has been suggested that defining the Hamiltonian such that the trace of the diabatic
matrix is minimised leads to more accurate results, but here we have not done this for
simplicity. --> Why? What do you expect by not implementing this? Can you add some information?

I really like this chapter