Notes Introduction to dynamicssimulations
Great! This page has turned out well indeed and is very easy to read and follow now!

For some reason, I can't get the by-line commenting feature to work anymore.

Line 29:, formerly "This takes the place of the `p` parameter ....": I've moved it a little and modified it, which hopefully makes even clearer how sim interacts with DifferentialEquations.

Line 74: I think it would be good to have a full list of the quantities and their names that can be outputed (or a link to that list) at this position. I'm not sure if it already has a manual entry, but there'll certainly be people who want to have a very special output, so linking to how to add a new quantity would be very helpful, I believe.