Update tully_scattering.md
This reads very well indeed. I have a few left-over comments:

line 3: roughly reproduce sounds odd, since it's not a quantitative assessment. 'We aim to reproduce'? Or something similar, perhaps.
line 31: what is the range of momentum values and what does the 9:2:50 mean? Do you select every two steps between 9 and 50? What does that mean for the trajectories? Do they successively get the assigned these values? What are the units?
- line 37: what is the timespan? Is that the entire time the trajectory will take or something else?
line 58: I thought we were comparing to Fig. 5 from the Tully paper? Which FSSH results do we reproduce accurately and how to you measure the accuracy? How do you know it will converge to the same results with more trajectories and why not do more here? (alternatively, say that there are some differences between us and whatever  you're comparing to and that the reason for it is the low number of trajectories we use in the example here.)

line 62 onwards: what are R1, T1, T2 in the legend of the plot? What does the plot tell us in general?
- line 44: you should mention somewhere that 2000 is the mass (and give the units for it)

I also made some modifications in the text, where I am not sure with all of them, so I added a question-mark to them.