Update nrpmd.md
This reads really well!
My comments:
- again, not all variables are declared that appear in the equations.
- line 102: what property is 5e3 and what is Dict(:X=>1.0)? Can we either add explanations or define them as separate variable so that becomes more obvious?
- line 103: similarly, what is 1/16 and what is size(sim)?
- line 115: what is vact?
(the reason why I'm insisting on things like this is that I would imagine that people who read the manual don't read it page after page, but rather look at the pages that are most relevant for them. It's also not easy to retain all information from reading the manual, so explaining all parameters would be very helpful. In this context, I really like lines 120-122!)
- line 140: excellent description! But how do :mean, append and sum combine the data?
- line 152: the plot needs x and y axes labels and a legend.